chat_test/requirements.txt
semantic-kernel>=1.0.0b8
openai>=1.25.0
python-dotenv>=1.0.1
azure-core>=1.30.1
azure-ai-ml>=1.15.0
aiohttp>=3.9.5
